chore: Move update_db_info to process_manager
