+++ backend/chatsky_ui/services/process_manager.py
@@ -12,19 +12,20 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 import socket
 
+from abc import ABC, abstractmethod
 from dotenv import load_dotenv
 from omegaconf import OmegaConf
 
 from chatsky_ui.core.config import settings
 from chatsky_ui.core.logger_config import get_logger
-from chatsky_ui.db.base import read_conf, read_logs
+from chatsky_ui.db.base import read_conf, read_logs, write_conf
 from chatsky_ui.schemas.preset import BuildPreset, RunPreset
 from chatsky_ui.schemas.process_status import Status
 from chatsky_ui.services.process import BuildProcess, RunProcess
 from chatsky_ui.utils.repo_manager import RepoManager
 
 
-class ProcessManager:
+class ProcessManager(ABC):
     """Base for build and run process managers."""
 
     def __init__(self):
@@ -88,7 +89,11 @@ class ProcessManager:
         for id_, process in self.processes.items():
             if await process.check_status() in [Status.ALIVE, Status.RUNNING]:
                 await self.stop(id_)
-                await process.update_db_info()
+        await self.update_db_info()
+
+    @abstractmethod
+    async def update_db_info(self):
+        raise NotImplementedError
 
     async def check_status(self, id_: int, *args, **kwargs) -> None:
         """Checks the status of the process with the given id by periodically checking status`
@@ -98,7 +103,7 @@ class ProcessManager:
         """
         process = self.processes[id_]
         while not process.to_be_terminated:
-            await process.update_db_info()  # check status and update db
+            await self.update_db_info()  # check status and update db
             process.logger.info("Status of process '%s': %s", process.id, process.status)
             if process.status in [
                 Status.NULL,
@@ -148,6 +153,18 @@ class ProcessManager:
 
         self.logger.info("Returning %s logs", len(logs))
         return logs[offset : offset + limit]
+
+    @staticmethod
+    def add_new_conf(conf: list, params: dict) -> list:  # TODO: rename conf everywhere to metadata/meta
+        for element in conf:
+            if element.id == params["id"]:  # type: ignore
+                for key, value in params.items():
+                    setattr(element, key, value)
+                break
+        else:
+            conf.append(params)
+
+        return conf
 
 
 class RunManager(ProcessManager):
@@ -204,6 +221,26 @@ class RunManager(ProcessManager):
 
     def get_port(self, run_id: int) -> Optional[int]:
         return self.processes[run_id].port
+
+    async def update_db_info(self) -> None:
+        # save current run info into runs_path
+        self.logger.debug("Updating db run info")
+        runs_conf = await read_conf(settings.runs_path)
+        for process in self.processes.values():
+            run_params = await process.get_full_info()
+            runs_conf = self.add_new_conf(runs_conf, run_params)  # type: ignore
+
+        await write_conf(runs_conf, settings.runs_path)
+
+        # save current run id into the correspoinding build in builds_path
+        builds_conf = await read_conf(settings.builds_path)
+        for build in builds_conf:
+            if build.id == run_params["build_id"]:  # type: ignore
+                if run_params["id"] not in build.run_ids:  # type: ignore
+                    build.run_ids.append(run_params["id"])  # type: ignore
+                    break
+
+        await write_conf(builds_conf, settings.builds_path)
 
 
 class BuildManager(ProcessManager):
@@ -272,7 +309,7 @@ class BuildManager(ProcessManager):
         """
         process = self.processes[id_]
         while not process.to_be_terminated:
-            await process.update_db_info()  # check status and update db
+            await self.update_db_info()  # check status and update db
             process.logger.info("Status of process '%s': %s", process.id, process.status)
             if process.status in [
                 Status.NULL,
@@ -296,3 +333,12 @@ class BuildManager(ProcessManager):
         Number of loglines returned is based on `offset` as the start line and limited by `limit` lines.
         """
         return await self.fetch_process_logs(build_id, offset, limit, settings.builds_path)
+
+    async def update_db_info(self) -> None:
+        """Saves current build info into builds_path"""
+        builds_conf = await read_conf(settings.builds_path)
+        for process in self.processes.values():
+            build_params = await process.get_full_info()
+            builds_conf = self.add_new_conf(builds_conf, build_params)  # type: ignore
+
+        await write_conf(builds_conf, settings.builds_path)
